% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AutodetectPython.R
\name{autodetectPython}
\alias{autodetectPython}
\title{Autodetects the settings for Windows}
\usage{
autodetectPython(pythonExePath = NULL)
}
\arguments{
\item{pythonExePath}{a string containing the path to "python.exe"
(e.g. "C:\\Python34\\python.exe").}
}
\value{
Returns a list containing the information necessary to
        connect to Python if a compatible Python, version was found,
        raises an error otherwise.
}
\description{
Autodetects the settings needed to connect to
             the python dll file (\strong{only Windows}).
}
\examples{
\dontrun{
  autodetectPython()
  autodetectPython("C:\\\\Python27\\\\python.exe")
}
}

