% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PySource.R
\name{BEGIN.Python}
\alias{BEGIN.Python}
\title{Execute Python interactively from within R}
\usage{
BEGIN.Python()
}
\value{
Returns the entered code as character, code lines which throw an
        exception are omitted.
}
\description{
The function BEGIN.Python starts an Python read-eval-print loop.
}
\details{
BEGIN.Python emulates the behavior of the Python terminal
         and therefore allows interactive Python code development
         from within R.
}
\note{
This won't work with RStudio because of a known
\href{https://support.rstudio.com/hc/communities/public/questions/206744317-readLines-Bug-?page=1#answer-206647788}{RStudio issue}.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
\dontrun{
code <-
BEGIN.Python()
import os
os.getcwd()
dir(os)
x = 3**3
for i in xrange(10):
    if (i > 5):
        print(i)

END.Python
## NOTE: BEGIN.Python returns the successfully executed code as character.
cat(code, sep="\\n")
pyGet0("x")
}
}

