% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Basics.R
\name{pyDir}
\alias{pyDir}
\title{Convenience function to call the Python function \strong{dir}}
\usage{
pyDir(objName = NULL)
}
\arguments{
\item{objName}{an optional string specifying the name of the Python object.}
}
\value{
Returns the list of names in the global scope, if no object name is
        provided, otherwise a list of valid attributes for the specified object.
}
\description{
A convenience function to call the Python function \strong{dir}.
}
\details{
The Python function dir is similar to the R function ls.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyDir()
pyDir("sys")
}

