% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyExec.R
\name{pyExec}
\alias{pyExec}
\title{Executes multiple lines of Python code from within R}
\usage{
pyExec(code)
}
\arguments{
\item{code}{a string of Python code to be executed in Python.}
}
\description{
The function pyExec allows to execute multiple lines of python
             code from within R.
}
\details{
Since pyExec can execute multiple lines, it is the obvious choice for defining
         Python functions or running small scripts where no return value is needed.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyExec('
print("The following line will not appear in the R terminal!")
"Hello" + " " + "R!"
print("NOTE: pyExecp would also show the line above!")
print("The following line will appear in the R terminal!")
print("Hello" + " " + "R!")
')
}

