% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyExec.R
\name{pyExecfile}
\alias{pyExecfile}
\title{Executes Python source file from within R}
\usage{
pyExecfile(filename)
}
\arguments{
\item{filename}{a character string giving the name or full path of the file to
be executed.}
}
\description{
The function pyExecfile calls the Python function \code{execfile}. which
             is the Python equivalent to the function \code{source} provided in R.
}
\details{
The function execfile is kind of the source of Python. Since it got omitted
         in Python 3 a replacement gets assigned following the suggestions from \cr
         \url{http://www.diveintopython3.net/porting-code-to-python-3-with-2to3.html}.
}
\examples{
\dontrun{
pyExecfile("myPythonScript.py")
}
}

