% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyObject.R
\name{pyFunction}
\alias{pyFunction}
\title{creates a virtual Python function}
\usage{
pyFunction(key)
}
\arguments{
\item{key}{a string specifying the name of a Python method/function.}
}
\description{
The function pyFunction creates a new object of type
             pyFunction based on a given key.
}
\details{
The function pyFunction makes it easy to create interfaces
         to Python functions.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
if ( pyIsConnected() ){
pySum <- pyFunction("sum")
pySum(1:3)
}
}

