% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyGet.R
\name{pyGet0}
\alias{pyGet0}
\title{Creates an R representation of an Python object}
\usage{
pyGet0(key)
}
\arguments{
\item{key}{a string specifying the name of a Python object.}
}
\value{
Returns the specified Python object converted into an R object if
        possible, else a virtual Python object.
}
\description{
The function pyGet0 gets Python objects by name.
}
\details{
Primitive data types like bool, int, long, float, str,
         bytes and unicode are returned as R objects. Python tuples, lists,
         dictionaries and other Python objects are returned as virtual
         Python objects.
}
\note{
pyGet0 never creates a new Python object.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
if ( pyIsConnected() ){
pyExec("import os")
os <- pyGet0("os")
os$getcwd()
os$sep
os$sep <- "Hello Python!"
pyExecp("os.sep")
}
}

