% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Basics.R
\name{pyHelp}
\alias{pyHelp}
\title{Convenience function to access the Python \strong{help} system}
\usage{
pyHelp(topic)
}
\arguments{
\item{topic}{a string specifying name or topic for which help is sought.}
}
\value{
Prints the help to the given string.
}
\description{
a convenience function to access the Python \strong{help} system.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyHelp("abs")
}

