% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PythonInR.R
\name{pyIsConnected}
\alias{pyIsConnected}
\title{checks if R is connected to Python}
\usage{
pyIsConnected()
}
\value{
Returns TRUE if R is connected to Python, FALSE otherwise.
}
\description{
Checks if R is connected to Python.
}
\examples{
pyIsConnected()
}

