% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PyExec.R
\name{pyPrint}
\alias{pyPrint}
\title{Convenience function to print a given Python object to the R terminal}
\usage{
pyPrint(objName)
}
\arguments{
\item{objName}{a character string to be evaluated in Python and printed
to the R terminal.}
}
\description{
Prints a given Python variable to the R terminal.
}
\details{
Internally it uses a combination of pyExec and print. Please note that
         the result of pyExecp("x") and pyPrint("x") often will be different,
         since pyExecp("x") is equivalent to typing x into the Python terminal
         whereas pyPrint("x") is equivalent to typing print(x) into the Python
         terminal.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyPrint("'Hello ' + 'R!'")
pyPrint("sys.version")
}

