% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PyOperators.R
\name{pyRef}
\alias{pyRef}
\title{get the location (reference) of an \code{PythonInR_Object}}
\usage{
pyRef(x)
}
\arguments{
\item{x}{an PythonInR_Object}
}
\description{
The function \code{pyRef} return the variable name
             of an virtual Python object.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
if ( pyIsConnected() ){
x <- pyObject("sys")
pyRef(x)
## [1] "sys"
}
}

