% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Basics.R
\name{pyType}
\alias{pyType}
\title{Convenience function to call the Python function \strong{type}}
\usage{
pyType(objName)
}
\arguments{
\item{objName}{a string specifying the name of the Python object.}
}
\value{
The type of the specified object as character on success, NULL otherwise.
}
\description{
Convenience function to call the Python function \strong{type}.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyExec("x = dict()")
pyType("x")
}

