% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PythonInR.R
\name{pyVersion}
\alias{pyVersion}
\title{is a convenience function to get sys.version from Python}
\usage{
pyVersion()
}
\value{
Returns a string containing the Python version and some compiler information.
}
\description{
A convenience function to get sys.version.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
pyVersion()
}

