% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PythonInR.R
\name{pyVersionInfo}
\alias{pyVersionInfo}
\title{is a convenience function to get sys.version_info from Python}
\usage{
pyVersionInfo()
}
\value{
Returns a list containing the Python version information.
}
\description{
A convenience function to get \code{sys.version}.
}
\examples{
\dontshow{PythonInR:::pyCranConnect()}
is.Python3 <- function() pyVersionInfo()[[1]] == 3L
}

